Muestra de dirección del viento como puntos cardinales en lugar de grados

+++ Practica 7 (Clima)/ClimaData.swift
@@ -36,5 +36,5 @@ struct Main: Codable {
 
 struct Wind: Codable {
     let speed: Double
-    let deg: Int
+    let deg: Double
 }
